+++ lib/internal/bootstrap/node.js
@@ -176,31 +176,6 @@ if (config.hasInspector) {
   internalBinding('inspector').registerAsyncHook(enable, disable);
 }
 
-// If the process is spawned with env NODE_CHANNEL_FD, it's probably
-// spawned by our child_process module, then initialize IPC.
-// This attaches some internal event listeners and creates:
-// process.send(), process.channel, process.connected,
-// process.disconnect()
-if (process.env.NODE_CHANNEL_FD) {
-  if (ownsProcessState) {
-    mainThreadSetup.setupChildProcessIpcChannel();
-  } else {
-    Object.defineProperty(process, 'channel', {
-      enumerable: false,
-      get: workerThreadSetup.unavailable('process.channel')
-    });
-
-    Object.defineProperty(process, 'connected', {
-      enumerable: false,
-      get: workerThreadSetup.unavailable('process.connected')
-    });
-
-    process.send = workerThreadSetup.unavailable('process.send()');
-    process.disconnect =
-      workerThreadSetup.unavailable('process.disconnect()');
-  }
-}
-
 const browserGlobals = !process._noBrowserGlobals;
 if (browserGlobals) {
   setupGlobalTimeouts();
