+++ lib/internal/bootstrap/pre_execution.js
@@ -2,6 +2,27 @@
 
 const { getOptionValue } = require('internal/options');
 
+function prepareMainThreadExecution() {
+  // If the process is spawned with env NODE_CHANNEL_FD, it's probably
+  // spawned by our child_process module, then initialize IPC.
+  // This attaches some internal event listeners and creates:
+  // process.send(), process.channel, process.connected,
+  // process.disconnect().
+  setupChildProcessIpcChannel();
+
+  // Load policy from disk and parse it.
+  initializePolicy();
+
+  // If this is a worker in cluster mode, start up the communication
+  // channel. This needs to be done before any user code gets executed
+  // (including preload modules).
+  initializeClusterIPC();
+
+  initializeDeprecations();
+  initializeESMLoader();
+  loadPreloadModules();
+}
+
 // In general deprecations are intialized wherever the APIs are implemented,
 // this is used to deprecate APIs implemented in C++ where the deprecation
 // utitlities are not easily accessible.
@@ -41,10 +62,22 @@ function initializeDeprecations() {
   }
 }
 
+function setupChildProcessIpcChannel() {
+  if (process.env.NODE_CHANNEL_FD) {
+    const assert = require('internal/assert');
+
+    const fd = parseInt(process.env.NODE_CHANNEL_FD, 10);
+    assert(fd >= 0);
+
+    // Make sure it's not accidentally inherited by child processes.
+    delete process.env.NODE_CHANNEL_FD;
+
+    require('child_process')._forkChild(fd);
+    assert(process.send);
+  }
+}
+
 function initializeClusterIPC() {
-  // If this is a worker in cluster mode, start up the communication
-  // channel. This needs to be done before any user code gets executed
-  // (including preload modules).
   if (process.argv[1] && process.env.NODE_UNIQUE_ID) {
     const cluster = require('cluster');
     cluster._setupWorker();
@@ -114,9 +147,8 @@ function loadPreloadModules() {
 }
 
 module.exports = {
+  prepareMainThreadExecution,
   initializeDeprecations,
-  initializeClusterIPC,
-  initializePolicy,
   initializeESMLoader,
   loadPreloadModules
 };
