+++ lib/internal/main/check_syntax.js
@@ -4,11 +4,7 @@
 // instead of actually running the file.
 
 const {
-  initializeDeprecations,
-  initializeClusterIPC,
-  initializePolicy,
-  initializeESMLoader,
-  loadPreloadModules
+  prepareMainThreadExecution
 } = require('internal/bootstrap/pre_execution');
 
 const {
@@ -22,11 +18,7 @@ const {
 } = require('internal/modules/cjs/helpers');
 
 // TODO(joyeecheung): not every one of these are necessary
-initializeDeprecations();
-initializeClusterIPC();
-initializePolicy();
-initializeESMLoader();
-loadPreloadModules();
+prepareMainThreadExecution();
 markBootstrapComplete();
 
 if (process.argv[1] && process.argv[1] !== '-') {
