+++ lib/internal/main/eval_stdin.js
@@ -3,11 +3,7 @@
 // Stdin is not a TTY, we will read it and execute it.
 
 const {
-  initializeDeprecations,
-  initializeClusterIPC,
-  initializePolicy,
-  initializeESMLoader,
-  loadPreloadModules
+  prepareMainThreadExecution
 } = require('internal/bootstrap/pre_execution');
 
 const {
@@ -15,11 +11,7 @@ const {
   readStdin
 } = require('internal/process/execution');
 
-initializeDeprecations();
-initializeClusterIPC();
-initializePolicy();
-initializeESMLoader();
-loadPreloadModules();
+prepareMainThreadExecution();
 markBootstrapComplete();
 
 readStdin((code) => {
