+++ lib/internal/main/eval_string.js
@@ -4,21 +4,13 @@
 // `--interactive`.
 
 const {
-  initializeDeprecations,
-  initializeClusterIPC,
-  initializePolicy,
-  initializeESMLoader,
-  loadPreloadModules
+  prepareMainThreadExecution
 } = require('internal/bootstrap/pre_execution');
 const { evalScript } = require('internal/process/execution');
 const { addBuiltinLibsToObject } = require('internal/modules/cjs/helpers');
 
 const source = require('internal/options').getOptionValue('--eval');
-initializeDeprecations();
-initializeClusterIPC();
-initializePolicy();
-initializeESMLoader();
-loadPreloadModules();
+prepareMainThreadExecution();
 addBuiltinLibsToObject(global);
 markBootstrapComplete();
 evalScript('[eval]', source, process._breakFirstLine);
