+++ lib/internal/main/run_main_module.js
@@ -1,18 +1,10 @@
 'use strict';
 
 const {
-  initializeDeprecations,
-  initializeClusterIPC,
-  initializePolicy,
-  initializeESMLoader,
-  loadPreloadModules
+  prepareMainThreadExecution
 } = require('internal/bootstrap/pre_execution');
 
-initializeDeprecations();
-initializeClusterIPC();
-initializePolicy();
-initializeESMLoader();
-loadPreloadModules();
+prepareMainThreadExecution();
 
 // Expand process.argv[1] into a full path.
 const path = require('path');
