+++ lib/internal/main/worker_thread.js
@@ -5,7 +5,6 @@
 
 const {
   initializeDeprecations,
-  initializeClusterIPC,
   initializeESMLoader,
   loadPreloadModules
 } = require('internal/bootstrap/pre_execution');
@@ -42,6 +41,26 @@ debug(`[${threadId}] is setting up worker child environment`);
 // Set up the message port and start listening
 const port = getEnvMessagePort();
 
+// If the main thread is spawned with env NODE_CHANNEL_FD, it's probably
+// spawned by our child_process module. In the work threads, mark the
+// related IPC properties as unavailable.
+if (process.env.NODE_CHANNEL_FD) {
+  const workerThreadSetup = require('internal/process/worker_thread_only');
+  Object.defineProperty(process, 'channel', {
+    enumerable: false,
+    get: workerThreadSetup.unavailable('process.channel')
+  });
+
+  Object.defineProperty(process, 'connected', {
+    enumerable: false,
+    get: workerThreadSetup.unavailable('process.connected')
+  });
+
+  process.send = workerThreadSetup.unavailable('process.send()');
+  process.disconnect =
+    workerThreadSetup.unavailable('process.disconnect()');
+}
+
 port.on('message', (message) => {
   if (message.type === LOAD_SCRIPT) {
     const {
@@ -57,7 +76,6 @@ port.on('message', (message) => {
       require('internal/process/policy').setup(manifestSrc, manifestURL);
     }
     initializeDeprecations();
-    initializeClusterIPC();
     initializeESMLoader();
     loadPreloadModules();
     publicWorker.parentPort = publicPort;
