+++ lib/internal/process/main_thread_only.js
@@ -152,22 +152,8 @@ function createSignalHandlers() {
   };
 }
 
-function setupChildProcessIpcChannel() {
-  const assert = require('internal/assert');
-
-  const fd = parseInt(process.env.NODE_CHANNEL_FD, 10);
-  assert(fd >= 0);
-
-  // Make sure it's not accidentally inherited by child processes.
-  delete process.env.NODE_CHANNEL_FD;
-
-  require('child_process')._forkChild(fd);
-  assert(process.send);
-}
-
 module.exports = {
   wrapProcessMethods,
   createSignalHandlers,
-  setupChildProcessIpcChannel,
   wrapPosixCredentialSetters
 };
